backend/requirements.txt
Django>=6.0,<6.1
djangorestframework>=3.17.0
djangorestframework-simplejwt>=5.5.0
django-cors-headers>=4.4.0
django-filter>=25.0
drf-spectacular>=0.29.0
Pillow>=10.0.0
gunicorn>=23.0.0
dj-database-url>=2.0.0
whitenoise>=6.7.0
django-jazzmin>=3.0.0
channels>=4.2.0
channels-redis>=4.3.0
celery>=5.4.0
redis>=5.0.0
